configuring webapck builder
