+++ public/index.html
@@ -1858,6 +1858,6 @@ Here is a card with filled data</textarea
       </section>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="js/bundle.js"></script>
   </body>
 </html>
